requirements.txt
# Core Dependencies
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
llama-cpp-python>=0.2.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0

# Data Processing
pypdf>=3.0.0
python-docx>=0.8.11
markdown>=3.4.0
tiktoken>=0.5.0

# Database & Storage
sqlitedict>=2.1.0

# Web Interface
streamlit>=1.28.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Security & Licensing
cryptography>=41.0.0
pyjwt>=2.8.0

# Configuration & Utilities
pyyaml>=6.0
click>=8.1.0
rich>=13.0.0
tqdm>=4.65.0

# Development & Testing (optional)
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# Additional dependencies for document processing
python-magic>=0.4.27
chardet>=5.2.0
